Move timezone specific variables to rsconf
