+++ runtime/glbl.c
@@ -58,6 +58,7 @@
 #include "queue.h"
 #include "dnscache.h"
 #include "parser.h"
+#include "timezones.h"
 
 /* some defaults */
 #ifndef DFLT_NETSTRM_DRVR
@@ -99,9 +100,6 @@ DEF_ATOMIC_HELPER_MUT(mutTerminateInputs);
 static int iFdSetSize = howmany(FD_SETSIZE, __NFDBITS) * sizeof (fd_mask); /* size of select() bitmask in bytes */
 #endif
 static uchar *SourceIPofLocalClient = NULL;	/* [ar] Source IP for local client to be used on multihomed host */
-
-tzinfo_t *tzinfos = NULL;
-static int ntzinfos;
 
 /* tables for interfacing with the v6 config system */
 static struct cnfparamdescr cnfparamdescr[] = {
@@ -182,16 +180,6 @@ static struct cnfparamblk paramblk =
 	{ CNFPARAMBLK_VERSION,
 	  sizeof(cnfparamdescr)/sizeof(struct cnfparamdescr),
 	  cnfparamdescr
-	};
-
-static struct cnfparamdescr timezonecnfparamdescr[] = {
-	{ "id", eCmdHdlrString, CNFPARAM_REQUIRED},
-	{ "offset", eCmdHdlrGetWord, CNFPARAM_REQUIRED }
-};
-static struct cnfparamblk timezonepblk =
-	{ CNFPARAMBLK_VERSION,
-	  sizeof(timezonecnfparamdescr)/sizeof(struct cnfparamdescr),
-	  timezonecnfparamdescr
 	};
 
 static struct cnfparamvals *cnfparamvals = NULL;
@@ -981,66 +969,6 @@ glblPrepCnf(void)
 	cnfparamvals = NULL;
 }
 
-
-static void
-freeTimezoneInfo(void)
-{
-	int i;
-	for(i = 0 ; i < ntzinfos ; ++i)
-		free(tzinfos[i].id);
-	free(tzinfos);
-	tzinfos = NULL;
-}
-
-static void
-displayTzinfos(void)
-{
-	int i;
-	if(!Debug)
-		return;
-	for(i = 0 ; i < ntzinfos ; ++i)
-		dbgprintf("tzinfo: '%s':%c%2.2d:%2.2d\n",
-			tzinfos[i].id, tzinfos[i].offsMode,
-			tzinfos[i].offsHour, tzinfos[i].offsMin);
-}
-
-
-/* Note: this function is NOT thread-safe!
- * This is currently not needed as used only during
- * initialization.
- */
-static rsRetVal
-addTimezoneInfo(uchar *tzid, char offsMode, int8_t offsHour, int8_t offsMin)
-{
-	DEFiRet;
-	tzinfo_t *newti;
-	CHKmalloc(newti = realloc(tzinfos, (ntzinfos+1)*sizeof(tzinfo_t)));
-	if((newti[ntzinfos].id = strdup((char*)tzid)) == NULL) {
-		free(newti);
-		DBGPRINTF("addTimezoneInfo: strdup failed with OOM\n");
-		ABORT_FINALIZE(RS_RET_OUT_OF_MEMORY);
-	}
-	newti[ntzinfos].offsMode = offsMode;
-	newti[ntzinfos].offsHour = offsHour;
-	newti[ntzinfos].offsMin = offsMin;
-	++ntzinfos, tzinfos = newti;
-finalize_it:
-	RETiRet;
-}
-
-
-static int
-bs_arrcmp_tzinfo(const void *s1, const void *s2)
-{
-	return strcmp((char*)s1, (char*)((tzinfo_t*)s2)->id);
-}
-/* returns matching timezone info or NULL if no entry exists */
-tzinfo_t*
-glblFindTimezoneInfo(char *id)
-{
-	return (tzinfo_t*) bsearch(id, tzinfos, ntzinfos, sizeof(tzinfo_t), bs_arrcmp_tzinfo);
-}
-
 /* handle the timezone() object. Each incarnation adds one additional
  * zone info to the global table of time zones.
  */
@@ -1051,81 +979,7 @@ bs_arrcmp_glblDbgFiles(const void *s1, const void *s2)
 	return strcmp((char*)s1, *(char**)s2);
 }
 
-void
-glblProcessTimezone(struct cnfobj *o)
-{
-	struct cnfparamvals *pvals;
-	uchar *id = NULL;
-	uchar *offset = NULL;
-	char offsMode;
-	int8_t offsHour;
-	int8_t offsMin;
-	int i;
 
-	pvals = nvlstGetParams(o->nvlst, &timezonepblk, NULL);
-	if(pvals == NULL) {
-		LogError(0, RS_RET_MISSING_CNFPARAMS, "error processing timezone "
-				"config parameters");
-		goto done;
-	}
-	if(Debug) {
-		dbgprintf("timezone param blk after glblProcessTimezone:\n");
-		cnfparamsPrint(&timezonepblk, pvals);
-	}
-
-	for(i = 0 ; i < timezonepblk.nParams ; ++i) {
-		if(!pvals[i].bUsed)
-			continue;
-		if(!strcmp(timezonepblk.descr[i].name, "id")) {
-			id = (uchar*) es_str2cstr(pvals[i].val.d.estr, NULL);
-		} else if(!strcmp(timezonepblk.descr[i].name, "offset")) {
-			offset = (uchar*) es_str2cstr(pvals[i].val.d.estr, NULL);
-		} else {
-			dbgprintf("glblProcessTimezone: program error, non-handled "
-			  "param '%s'\n", timezonepblk.descr[i].name);
-		}
-	}
-
-	/* note: the following two checks for NULL are not strictly necessary
-	 * as these are required parameters for the config block. But we keep
-	 * them to make the clang static analyzer happy, which also helps
-	 * guard against logic errors.
-	 */
-	if(offset == NULL) {
-		parser_errmsg("offset parameter missing (logic error?), timezone config ignored");
-		goto done;
-	}
-	if(id == NULL) {
-		parser_errmsg("id parameter missing (logic error?), timezone config ignored");
-		goto done;
-	}
-
-	if(   strlen((char*)offset) != 6
-	   || !(offset[0] == '-' || offset[0] == '+')
-	   || !(isdigit(offset[1]) && isdigit(offset[2]))
-	   || offset[3] != ':'
-	   || !(isdigit(offset[4]) && isdigit(offset[5]))
-	  ) {
-		parser_errmsg("timezone offset has invalid format. Must be +/-hh:mm, e.g. \"-07:00\".");
-		goto done;
-	}
-
-	offsHour = (offset[1] - '0') * 10 + offset[2] - '0';
-	offsMin  = (offset[4] - '0') * 10 + offset[5] - '0';
-	offsMode = offset[0];
-
-	if(offsHour > 12 || offsMin > 59) {
-		parser_errmsg("timezone offset outside of supported range (hours 0..12, minutes 0..59)");
-		goto done;
-	}
-
-	addTimezoneInfo(id, offsMode, offsHour, offsMin);
-
-done:
-	cnfparamvalsDestruct(pvals, &timezonepblk);
-	free(id);
-	free(offset);
-}
 
 /* handle a global config object. Note that multiple global config statements
  * are permitted (because of plugin support), so once we got a param block,
@@ -1221,15 +1075,6 @@ glblDestructMainqCnfObj(void)
 	}
 }
 
-/* comparison function for qsort() and string array compare
- * this is for the string lookup table type
- */
-static int
-qs_arrcmp_tzinfo(const void *s1, const void *s2)
-{
-	return strcmp(((tzinfo_t*)s1)->id, ((tzinfo_t*)s2)->id);
-}
-
 static int
 qs_arrcmp_glblDbgFiles(const void *s1, const void *s2)
 {
@@ -1288,11 +1133,9 @@ glblDoneLoadCnf(void)
 	DEFiRet;
 	CHKiRet(objUse(net, CORE_COMPONENT));
 
-	if(ntzinfos > 0) {
-		qsort(tzinfos, ntzinfos, sizeof(tzinfo_t), qs_arrcmp_tzinfo);
-	}
-	DBGPRINTF("Timezone information table (%d entries):\n", ntzinfos);
-	displayTzinfos();
+	sortTimezones(loadConf);
+	DBGPRINTF("Timezone information table (%d entries):\n", loadConf->timezones.ntzinfos);
+	displayTimezones(loadConf);
 
 	if(cnfparamvals == NULL)
 		goto finalize_it;
@@ -1575,7 +1418,6 @@ BEGINObjClassExit(glbl, OBJ_IS_CORE_MODULE) /* class, version */
 	free(LocalHostName);
 	free(LocalHostNameOverride);
 	free(LocalFQDNName);
-	freeTimezoneInfo();
 	objRelease(prop, CORE_COMPONENT);
 	if(propLocalHostNameToDelete != NULL)
 		prop.Destruct(&propLocalHostNameToDelete);
